Comment out the prod db setting and added more to the models
This is the beginning of having models as part of our project.
I also commented out the prod db setting because if you don't have the
driver for it you can't migrate even for local. Lastly I added *.pyc to
the .gitignore because these should not be in the repo.