made some convience changes and clean some things up
there is now a header so that no header code for the site needs to be
duplicated. also there where some simple things cleaned up in the code
the password is now hidden in firefox and the delete button does not
show up on an add. but saddly the user jail is still off because I
could not set up the login page for exeption this should be looked at
sometime this week hopefully. otherwise carry on.